pyproject.toml
[project]
name = "easybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msgpack>=1.1.2",
    "pyzmq>=27.1.0",
    "sanic>=25.12.0",
]
